pyproject.toml
[tool.poetry]
name = "todo-bot"
version = "0.1.0"
description = ""
authors = ["Vitaliy Angolenko <v.angolenko@deepvision.team>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = "^20.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
bot = "bot:main"